Fixing symlinks so windows can compile properly in github CI.
When running in github CI, symlinks are proper symlinks and the main.yml file replaces the symlinks with a copy of the linked file. 
So in the SConstruct, if the GITHUB_REPOSITORY env var exists, we disable the win32 specific code that translates the symlink text file created in windows to the symlink target.